% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{ask_col_mean}
\alias{ask_col_mean}
\title{Ask to Take Mean of a Column}
\usage{
ask_col_mean(dataframe)
}
\arguments{
\item{dataframe}{A data.frame. Should have numeric data columns to be
summarised.}
}
\value{
Nothing.
}
\description{
Ask to Take Mean of a Column
}
\details{
The suer will be prompted to choose a numeric column from the
data.frame object provided in 'dataframe'.
}
\examples{
ask_col_mean(mtcars)

}
