% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosethis-package.R
\docType{package}
\name{choosethis-package}
\alias{choosethis}
\alias{choosethis-package}
\title{choosethis: Demos of 'cli' Clickable Links}
\description{
Two sketches: play a text adventure in the console by following the prompts, and perform a data summary using a clickabel interface. Options are presented to the user as clickable links in supported terminals, such as RStudio's console.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
