% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{choosethis_opts}
\alias{choosethis_opts}
\title{Set Options for 'choosethis' Output}
\usage{
choosethis_opts(show_links = cli::ansi_has_hyperlink_support())
}
\arguments{
\item{show_links}{Logical. Should options be presented as clickable links
using 'cli'? Defaults to automatic detection by
\link[cli:style_hyperlink]{cli::ansi_has_hyperlink_support}.}
}
\value{
A list.
}
\description{
Set Options for 'choosethis' Output
}
\examples{
choosethis_opts(show_links = FALSE)

}
