% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{get_mean}
\alias{get_mean}
\title{Get the Mean}
\usage{
get_mean(dat, col)
}
\arguments{
\item{dat}{A data.frame. A data frame with numeric columns.}

\item{col}{Character. The name of a numeric column from the 'dat' data.frame.}
}
\value{
Numeric. The mean of the numeric column 'col' from data.frame 'dat'.
}
\description{
Evaluate the mean of a named column. NAs are removed automatically.
}
\examples{
get_mean(mtcars, "drat")

}
