% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosethis-package.R
\docType{package}
\name{choosethis-package}
\alias{choosethis}
\alias{choosethis-package}
\title{choosethis: Choose Your Own Adventure}
\description{
Play a 'Choose Your Own Adventure' game in the console by following the prompts. Options are presented as clickable links in certain environments, such as RStudio.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
